backend/requirements.txt
Flask
pyspark
pyhive
pandas
kafka-python
flask-cors
thrift
thrift_sasl 
# sasl
# pyhive[hive]